Change double data type to float and improve documentation

+++ src/game_system.cpp
@@ -4,20 +4,21 @@
 /// Array of each decrease rate for each difficulty.
 const float DECREASE_RATES[] = {0.05, 0.07, 0.09, 0.11};
 
+//Euler's constant
 const float e = 2.7182818284;
 
-/// Initial leds turning off time
+/// Initial leds turning off time in milliseconds
 const unsigned int INITIAL_T2 = 3000;
 
-/// Initial time Player have to finish the level on easiest difficulty
+/// Initial time Player have to finish the level on easiest difficulty in milliseconds
 const unsigned int INITIAL_T3 = 10000;
 
 /// @brief Apply a formula that follow an'exponential decrese of the time
-/// @param initialValue 
-/// @param decreaseRatio 
-/// @param level 
+/// @param initialValue: starting value 
+/// @param decreaseRatio: the greater, the fastest the number goes down 
+/// @param level: starting from 0, more level means less time, less level more time
 /// @return 
-double ApplyDecreasingFormula(float initialValue, float decreaseRatio, int level) {
+float ApplyDecreasingFormula(float initialValue, float decreaseRatio, int level) {
   return initialValue * pow(e, -decreaseRatio * level);
 }
 
@@ -25,7 +26,7 @@ double ApplyDecreasingFormula(float initialValue, float decreaseRatio, int level
 /// @param level: current level.
 /// @param difficulty: current difficulty.
 /// @return the new T2.
-double CalculateT2(int level, int difficulty) {
+float CalculateT2(int level, int difficulty) {
   return ApplyDecreasingFormula(INITIAL_T2, DECREASE_RATES[difficulty], level);
 }
 
@@ -33,6 +34,6 @@ double CalculateT2(int level, int difficulty) {
 /// @param level: current level.
 /// @param difficulty: current difficulty.
 /// @return the new T3.
-double CalculateT3(int level, int difficulty) {
+float CalculateT3(int level, int difficulty) {
   return ApplyDecreasingFormula(INITIAL_T3, DECREASE_RATES[difficulty], level);
 }
